자유게시판(예신) 수정 기능
