+++ src/main/java/com/beanions/common/dao/user/board/FreeBoardMapper.java
@@ -1,9 +1,7 @@
 package com.beanions.common.dao.user.board;
 
-import com.beanions.common.dto.MainCategoryDTO;
+import com.beanions.common.dto.*;
 import com.beanions.common.dto.PostAndMemberDTO;
-import com.beanions.common.dto.PostAndMemberDTO;
-import com.beanions.common.dto.SearchDTO;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
@@ -15,9 +13,12 @@ public interface FreeBoardMapper {
     int count(SearchDTO params);
 
 
-    List<PostAndMemberDTO> yesinDetail(String id);
+    List<PostAndMemberDTO> yesinDetail(int id);
 
     void postRegist(PostAndMemberDTO newPost);
 
     List<MainCategoryDTO> findMainCategory();
+
+    void postModify(PostDTO postDTO);
+
 }
