+++ src/main/java/com/beanions/user/controller/board/NoticeController.java
@@ -38,25 +38,4 @@ public class NoticeController {
 
         return "user/board/noticeboarddetail";
     }
-
-    @GetMapping("/modify")
-    public String modifyPost(@RequestParam("id") String id, Model model){
-
-
-        List<PostDTO> modify = noticeService.selectNotice(id);
-        model.addAttribute("modify", modify);
-
-        System.out.println(modify);
-
-        return "user/board/modify";
-    }
-
-    @PostMapping("/modify")
-    public String modifyPost(PostDTO postDTO, RedirectAttributes rttr){
-
-       noticeService.modifyPost(postDTO);
-        rttr.addFlashAttribute("successMessage", "수정 성공..");
-
-        return "redirect:/notice/board/board";
-    }
 }
