+++ src/main/java/com/beanions/user/service/board/FreeBoardService.java
@@ -3,10 +3,12 @@ package com.beanions.user.service.board;
 import com.beanions.common.dao.user.board.FreeBoardMapper;
 import com.beanions.common.dto.MainCategoryDTO;
 import com.beanions.common.dto.PostAndMemberDTO;
+import com.beanions.common.dto.PostDTO;
 import com.beanions.common.dto.SearchDTO;
 import com.beanions.common.paging.Pagination;
 import com.beanions.common.paging.PagingResponse;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Collections;
 import java.util.List;
@@ -46,7 +48,9 @@ public class FreeBoardService {
 
     public List<PostAndMemberDTO> yesinDetail(String id) {
 
-        return freeBoardMapper.yesinDetail(id);
+        int code = Integer.parseInt(id);
+
+        return freeBoardMapper.yesinDetail(code);
     }
 
     public void postRegist(PostAndMemberDTO newPost) {
@@ -56,5 +60,10 @@ public class FreeBoardService {
 
     public List<MainCategoryDTO> findMainCategory() {
         return freeBoardMapper.findMainCategory();
+    }
+
+    @Transactional
+    public void modifyPost(PostDTO postDTO) {
+        freeBoardMapper.postModify(postDTO);
     }
 }
