+++ src/main/java/com/beanions/user/service/board/NoticeService.java
@@ -28,10 +28,4 @@ public class NoticeService {
         int code = Integer.parseInt(id);
         return noticeMapper.noticeSelectOneDetail(code);
     }
-    
-    @Transactional
-    public void modifyPost(PostDTO id) {
-
-        noticeMapper.postModify(id);
-    }
 }
