+++ src/main/resources/mapper/userBoard/FreeBoardMapper.xml
@@ -56,17 +56,23 @@
 
     <select id="yesinDetail" resultMap="PostAndMemberResultMap">
         SELECT
-        A.POST_CODE
-        , A.POST_TITLE
-        , A.MEMBER_CODE
-        , A.POST_DATE
-        , A.MAIN_CATEGORY
-        , A.SUB_CATEGORY
-        , A.POST_CONTEXT
-        , B.MEMBER_CODE
-        , B.NICKNAME
-        FROM POST A
-        LEFT JOIN MEMBERS B ON A.MEMBER_CODE = B.MEMBER_CODE
+        POST_CODE,
+        MEMBER_CODE,
+        POST_TITLE,
+        POST_CONTEXT,
+        POST_DATE,
+        MAIN_CATEGORY,
+        SUB_CATEGORY
+        FROM POST
+        WHERE SUB_CATEGORY = '예신' AND POST_CODE = #{ id }
+        ORDER BY POST_CODE
     </select>
 
+    <update id="postModify" parameterType="com.beanions.common.dto.PostDTO">
+        UPDATE POST
+           SET
+               POST_TITLE = #{ postTitle },
+               POST_CONTEXT = #{ postContext }
+         WHERE POST_CODE = #{ postCode }
+    </update>
 </mapper>
