+++ src/main/resources/mapper/userBoard/NoticeMapper.xml
@@ -52,13 +52,4 @@
         WHERE MAIN_CATEGORY = '공지사항' AND POST_CODE = #{ id }
         ORDER BY POST_CODE
     </select>
-
-    <update id="postModify" parameterType="com.beanions.common.dto.PostDTO">
-        UPDATE POST
-           SET
-               POST_TITLE = #{ postTitle },
-               POST_CONTEXT = #{ postContext }
-         WHERE POST_CODE = #{ postCode }
-    </update>
-
 </mapper>
