+++ src/main/resources/templates/user/board/freeboardmodify.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+<h1>게시물 수정</h1>
+<form action="/user/board/modify" method="post">
+    <div th:each="post : ${ modify }">
+        <input type="hidden" th:value="${post.postCode}" name="postCode">
+        <input th:value="${post.postTitle}" name="postTitle"><br>
+        <input th:value="${post.postContext}" name="postContext">
+        <button type="submit">수정하기</button>
+        <button type="button" th:onclick="|location.href='@{/user/board/yesinDetail(id=${post.postCode})}'|">이전으로</button>
+    </div>
+</form>
+</body>
+</html>
