--- src/main/resources/templates/user/board/modify.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-
-<h1>여기</h1>
-<form action="/notice/modify" method="post">
-    <div th:each="post : ${ modify }">
-        <input th:value="${post.postTitle}" name="postTitle">
-        <input th:value="${post.postContext}" name="postContext">
-    </div>
-    <button type="submit" value="수정하기"></button>
-</form>
-</body>
-</html>
